% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupseq.R
\docType{package}
\name{groupseq.pkg}
\alias{groupseq.pkg}
\title{Group sequential design probabilities - with graphical user interface.}
\description{
Computes probabilities related to group sequential designs for
normally distributed test statistics. Enables to derive critical
boundaries, power, drift, and confidence intervals of such designs.
Supports the alpha spending approach by Lan-DeMets as well as the
conditional rejection probability principle by Müller and Schäfer.
}
\author{
Roman Pahl, \email{roman.pahl@gmail.com}
}
