% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{gui}
\alias{gui}
\title{Create graphical user interface.}
\usage{
gui(root)
}
\value{
Invoked for its side effects.
}
\description{
This function builds the main GUI that appears when \link{GroupSeq}
is started.
}
\keyword{internal}
