\name{guiInputTask4}

\alias{guiInputTask4}

\title{ input for calculating Task 4 }

\description{
This function creates a GUI to get input for calculating Task 4
}

\usage{
guiInputTask4()
}

\author{ Roman Pahl }

\note{This function is meant to be called in context of using \code{\link{groupseq}}\cr
      It won´t work on its own!}

\seealso{ \code{\link{groupseq}}, \code{\link{guiInputTask1}} 
          \code{\link{guiInputTask2}}, \code{\link{guiInputTask3}}}

\keyword{methods}
\keyword{IO}