% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential.R
\name{seqDesign}
\alias{seqDesign}
\title{Sequential design probabilities}
\usage{
seqDesign(
  crit,
  mu = 0,
  sd = 1,
  tau = seq_along(crit)/length(crit),
  futStop = rep(-Inf, length(crit)),
  n = 1,
  ...
)
}
\arguments{
\item{crit}{\code{numeric} vector of critical limits for stopping the study with
rejection of the null hypothesis. The length of 'crit' determines the
number of stages of the design.}

\item{mu}{\code{numeric} the mean (e.g. \code{mu == 0} under H0). Under H1 this
is also known as the drift parameter.}

\item{sd}{\code{numeric} standard deviation (or vector of sd)}

\item{tau}{\code{numeric} vector of monotonely increasing information times.}

\item{futStop}{\code{numeric} vector of futility stopping bounds (optional).}

\item{n}{\code{numeric} sample size, which only has an effect if \code{mu != 0}.
Useful to compute expected sample size under H1.}

\item{...}{further arguments passed to \code{seqProb}}
}
\value{
result is a \code{list} with the following components:
\itemize{
\item \code{exitP} the alpha spent at each stage. The cumulative sum of this will
provides the cumulative exit probability up to each stage, that is, the
probability of having stopped due to exceeding a bound.
\item \code{undecided} probability of continuing the study. In case of the last
stage, this means stopping without declaring significance.
\item \code{cumFutStop} cumulative futility stop probabilities.
\item \code{nk} planned sample size per stage.
\item \code{kExp} expected sample size per stage - the sum over all values will
provide the total expected sample size.
}
}
\description{
Computes sequential probabilities, for each stage of a given
one-sided sequential design with specified critical limits and optional
futility stops (stopping bounds). When applying the design in practice, at
each stage, all data collected up to this point is used to calculate a
combined test statistic. Then, if the test statistic is larger than the
critical limit at the specific stage, the study is stopped and the null
hypothesis rejected. If futility stops have been specified and the test
statistic falls below them at any stage, the study is also stopped due to
futility while the null hypothesis in this case cannot be discarded.
}
\examples{
# 1-stage
# #######
alpha <- 0.05
crit <- qnorm(1-alpha)
seqDesign(crit)$p$exit                          # type1: 0.05 == alpha
seqDesign(crit, mu=0.3)$p$exit                  # power: 9\%
# Increase sample size
seqDesign(crit, mu=0.3, n=69)$p$exit            # power: 80.1\%

# 2-stage
# #######
crit2 <- rep(crit, 2)

# Using two stages with same critical bounds will exceed alpha
cumsum(seqDesign(crit2, mu=0)$p$exit)           # cumulative type1: > 0.08
cumsum(seqDesign(crit2, mu=0, n=69)$p$exit)     # type1 does not depend on n

# Try Bonferroni:
bonf <- rep(qnorm(1-alpha/2), 2)
cumsum(seqDesign(bonf)$p$exit)                  # cumulative type1: 0.0415...
sum(seqDesign(bonf, mu=0.3, n=69)$p$exit)       # cumulative power: 73.1\%

# Bonferroni did not use all alpha, so there is room to improve power, for
# example using Pocock like boundaries:
pocock <- c(1.8662, 1.8848)
sum(seqDesign(pocock)$p$exit)                   # cumulative type1: 0.05
sum(seqDesign(pocock, mu=.3, n=69)$p$exit)      # power: 75.8\%

# Still the one stage power will never be reached, which is the price to pay
# for the two potential tests. On the other hand, the expected sample size is
# considerably smaller in comparison with the 1-stage design.
sum(seqDesign(pocock, mu=.3, n=69)$n$nkExp)     # expected sample size: 53

# Increasing max sample size to 98, yields expected n = 69 and power 87.8\%
res <- seqDesign(pocock, mu=.3, n=98)
lapply(unlist(res, recursive=FALSE)[c("n.nkExp", "p.exit")], sum)
}
