% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupseq.R
\name{start_gui}
\alias{start_gui}
\title{Start GroupSeq}
\usage{
start_gui(legacy = FALSE)
}
\arguments{
\item{legacy}{\code{logical} if \code{TRUE}, starts legacy GroupSeq version < 2.}
}
\description{
Starts the graphical user interface, optionally the legacy
version of GroupSeq prior to version 2.
}
