% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verify_arg}
\alias{verify_arg}
\title{Verify function arguments}
\usage{
verify_arg(x, type = "character", len = 1L, throw = TRUE)
}
\arguments{
\item{x}{object to verify}

\item{type}{\code{character} expected class or type}

\item{len}{\code{integer} expected length of argument}

\item{throw}{\code{logical} if TRUE, an error is signaled if verification fails,
otherwise just \code{FALSE} is returned.}
}
\value{
returns invisibly \code{TRUE} if verification succeeds
}
\description{
Enables shortened arg checking
}
